+++ lnd-grpc-client/lnd-grpc-client-autoconfigure/src/main/java/org/tbk/lightning/lnd/grpc/config/LndClientAutoConfiguration.java
@@ -9,6 +9,26 @@ import lombok.SneakyThrows;
 import org.lightningj.lnd.wrapper.AsynchronousLndAPI;
 import org.lightningj.lnd.wrapper.MacaroonContext;
 import org.lightningj.lnd.wrapper.SynchronousLndAPI;
+import org.lightningj.lnd.wrapper.autopilot.AsynchronousAutopilotAPI;
+import org.lightningj.lnd.wrapper.autopilot.SynchronousAutopilotAPI;
+import org.lightningj.lnd.wrapper.chainnotifier.AsynchronousChainNotifierAPI;
+import org.lightningj.lnd.wrapper.chainnotifier.SynchronousChainNotifierAPI;
+import org.lightningj.lnd.wrapper.invoices.AsynchronousInvoicesAPI;
+import org.lightningj.lnd.wrapper.invoices.SynchronousInvoicesAPI;
+import org.lightningj.lnd.wrapper.router.AsynchronousRouterAPI;
+import org.lightningj.lnd.wrapper.router.SynchronousRouterAPI;
+import org.lightningj.lnd.wrapper.signer.AsynchronousSignerAPI;
+import org.lightningj.lnd.wrapper.signer.SynchronousSignerAPI;
+import org.lightningj.lnd.wrapper.verrpc.AsynchronousVersionerAPI;
+import org.lightningj.lnd.wrapper.verrpc.SynchronousVersionerAPI;
+import org.lightningj.lnd.wrapper.walletkit.AsynchronousWalletKitAPI;
+import org.lightningj.lnd.wrapper.walletkit.SynchronousWalletKitAPI;
+import org.lightningj.lnd.wrapper.walletunlocker.AsynchronousWalletUnlockerAPI;
+import org.lightningj.lnd.wrapper.walletunlocker.SynchronousWalletUnlockerAPI;
+import org.lightningj.lnd.wrapper.watchtower.AsynchronousWatchtowerAPI;
+import org.lightningj.lnd.wrapper.watchtower.SynchronousWatchtowerAPI;
+import org.lightningj.lnd.wrapper.wtclient.AsynchronousWatchtowerClientAPI;
+import org.lightningj.lnd.wrapper.wtclient.SynchronousWatchtowerClientAPI;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
@@ -95,10 +115,14 @@ public class LndClientAutoConfiguration {
                 .build();
     }
 
-    @Bean(name = "synchronousLndClient", destroyMethod = "close")
+    // ----------- API beans ------------
+    // see https://www.lightningj.org/#_available_apis for all available LightningJ APIs
+
+    // ----------- Lightning API beans
+    @Bean(name = "synchronousLndAPI", destroyMethod = "close")
     @ConditionalOnMissingBean
     @ConditionalOnBean(LndRpcConfig.class)
-    public SynchronousLndAPI synchronousLndClient(LndRpcConfig rpcConfig) {
+    public SynchronousLndAPI synchronousLndAPI(LndRpcConfig rpcConfig) {
         return new SynchronousLndAPI(
                 rpcConfig.getRpchost(),
                 rpcConfig.getRpcport(),
@@ -106,10 +130,10 @@ public class LndClientAutoConfiguration {
                 rpcConfig.getMacaroonContext());
     }
 
-    @Bean(name = "lndClient", destroyMethod = "close")
+    @Bean(name = "lndAPI", destroyMethod = "close")
     @ConditionalOnMissingBean
     @ConditionalOnBean(LndRpcConfig.class)
-    public AsynchronousLndAPI lndClient(LndRpcConfig rpcConfig) {
+    public AsynchronousLndAPI lndAPI(LndRpcConfig rpcConfig) {
         return new AsynchronousLndAPI(
                 rpcConfig.getRpchost(),
                 rpcConfig.getRpcport(),
@@ -117,4 +141,233 @@ public class LndClientAutoConfiguration {
                 rpcConfig.getMacaroonContext());
     }
 
+    // ----------- Wallet API beans
+    @Bean(name = "synchronousLndWalletUnlockerAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousWalletUnlockerAPI synchronousLndWalletUnlockerAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousWalletUnlockerAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndWalletUnlockerAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousWalletUnlockerAPI lndWalletUnlockerAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousWalletUnlockerAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- Autopilot API beans
+    @Bean(name = "synchronousLndAutopilotAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousAutopilotAPI synchronousLndAutopilotAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousAutopilotAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndAutopilotAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousAutopilotAPI lndAutopilotAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousAutopilotAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- ChainNotifier API beans
+    @Bean(name = "synchronousLndChainNotifierAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousChainNotifierAPI synchronousLndChainNotifierAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousChainNotifierAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndChainNotifierAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousChainNotifierAPI lndChainNotifierAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousChainNotifierAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- Invoices API beans
+    @Bean(name = "synchronousLndInvoiceAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousInvoicesAPI synchronousLndInvoiceAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousInvoicesAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndInvoiceAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousInvoicesAPI lndInvoiceAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousInvoicesAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- Router API beans
+    @Bean(name = "synchronousLndRouterAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousRouterAPI synchronousLndRouterAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousRouterAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndRouterAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousRouterAPI lndRouterAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousRouterAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- Signer API beans
+    @Bean(name = "synchronousLndSignerAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousSignerAPI synchronousLndSignerAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousSignerAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndSignerAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousSignerAPI lndSignerAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousSignerAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- WalletKit API beans
+    @Bean(name = "synchronousLndWalletKitAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousWalletKitAPI synchronousLndWalletKitAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousWalletKitAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndWalletKitAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousWalletKitAPI lndWalletKitAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousWalletKitAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- Watchtower API beans
+    @Bean(name = "synchronousLndWatchtowerAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousWatchtowerAPI synchronousLndWatchtowerAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousWatchtowerAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndWatchtowerAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousWatchtowerAPI lndWatchtowerAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousWatchtowerAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- Watchtower Client API beans
+    @Bean(name = "synchronousLndWatchtowerClientAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousWatchtowerClientAPI synchronousLndWatchtowerClientAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousWatchtowerClientAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndWatchtowerClientAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousWatchtowerClientAPI lndWatchtowerClientAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousWatchtowerClientAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    // ----------- Versioning API beans
+    @Bean(name = "synchronousLndVersionerAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public SynchronousVersionerAPI synchronousLndVersionerAPI(LndRpcConfig rpcConfig) {
+        return new SynchronousVersionerAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
+
+    @Bean(name = "lndVersionerAPI", destroyMethod = "close")
+    @ConditionalOnMissingBean
+    @ConditionalOnBean(LndRpcConfig.class)
+    public AsynchronousVersionerAPI lndVersionerAPI(LndRpcConfig rpcConfig) {
+        return new AsynchronousVersionerAPI(
+                rpcConfig.getRpchost(),
+                rpcConfig.getRpcport(),
+                rpcConfig.getSslContext(),
+                rpcConfig.getMacaroonContext());
+    }
 }