+++ lnd-grpc-client/lnd-grpc-client-autoconfigure/src/test/java/org/tbk/lightning/lnd/grpc/config/LndClientAutoConfigurationTest.java
@@ -1,10 +1,23 @@
 package org.tbk.lightning.lnd.grpc.config;
 
 import com.google.common.base.Throwables;
+import com.google.common.collect.ImmutableMap;
 import org.junit.jupiter.api.Test;
 import org.lightningj.lnd.wrapper.AsynchronousLndAPI;
+import org.lightningj.lnd.wrapper.autopilot.AsynchronousAutopilotAPI;
+import org.lightningj.lnd.wrapper.chainnotifier.AsynchronousChainNotifierAPI;
+import org.lightningj.lnd.wrapper.invoices.AsynchronousInvoicesAPI;
+import org.lightningj.lnd.wrapper.router.AsynchronousRouterAPI;
+import org.lightningj.lnd.wrapper.signer.AsynchronousSignerAPI;
+import org.lightningj.lnd.wrapper.verrpc.AsynchronousVersionerAPI;
+import org.lightningj.lnd.wrapper.walletkit.AsynchronousWalletKitAPI;
+import org.lightningj.lnd.wrapper.walletunlocker.AsynchronousWalletUnlockerAPI;
+import org.lightningj.lnd.wrapper.watchtower.AsynchronousWatchtowerAPI;
+import org.lightningj.lnd.wrapper.wtclient.AsynchronousWatchtowerClientAPI;
 import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
+
+import java.util.Map;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
@@ -25,8 +38,24 @@ public class LndClientAutoConfigurationTest {
                         "org.tbk.lightning.lnd.grpc.certFilePath=src/test/resources/lnd/tls-test.cert"
                 )
                 .run(context -> {
-                    assertThat(context.containsBean("lndClient"), is(true));
-                    assertThat(context.getBean(AsynchronousLndAPI.class), is(notNullValue()));
+                    Map<String, Class<?>> beanNamesAndClasses = ImmutableMap.<String, Class<?>>builder()
+                            .put("lndAPI", AsynchronousLndAPI.class)
+                            .put("lndWalletUnlockerAPI", AsynchronousWalletUnlockerAPI.class)
+                            .put("lndAutopilotAPI", AsynchronousAutopilotAPI.class)
+                            .put("lndChainNotifierAPI", AsynchronousChainNotifierAPI.class)
+                            .put("lndInvoiceAPI", AsynchronousInvoicesAPI.class)
+                            .put("lndRouterAPI", AsynchronousRouterAPI.class)
+                            .put("lndSignerAPI", AsynchronousSignerAPI.class)
+                            .put("lndWalletKitAPI", AsynchronousWalletKitAPI.class)
+                            .put("lndWatchtowerAPI", AsynchronousWatchtowerAPI.class)
+                            .put("lndWatchtowerClientAPI", AsynchronousWatchtowerClientAPI.class)
+                            .put("lndVersionerAPI", AsynchronousVersionerAPI.class)
+                            .build();
+
+                    beanNamesAndClasses.forEach((name, clazz) -> {
+                        assertThat(context.containsBean(name), is(true));
+                        assertThat(context.getBean(clazz), is(notNullValue()));
+                    });
                 });
     }
 
@@ -37,7 +66,7 @@ public class LndClientAutoConfigurationTest {
                         "org.tbk.lightning.lnd.grpc.enabled=false"
                 )
                 .run(context -> {
-                    assertThat(context.containsBean("lndClient"), is(false));
+                    assertThat(context.containsBean("lndAPI"), is(false));
                     assertThrows(NoSuchBeanDefinitionException.class, () -> context.getBean(AsynchronousLndAPI.class));
                 });
     }